[Fixed] 별점 기능 추가
TotalRate 더블형으로 변수 수정
rateAvg 변수 추가

StoreService 코드 추가
rateAvg 계산 코드 추가

StoreResponseDto ...
getAllStore, getStore ..
rateAvg 인자추가

+++ src/main/java/com/example/Fooding/review/entity/Review.java
@@ -27,21 +27,21 @@ public class Review {
     private String reviewContent;
 
     @Column(name = "rate")
-    private Long rate;
+    private Double rate;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "store_id", nullable = false)
     private Store store;
 
     @Builder
-    public Review(Long makerId, String reviewContent, Long rate, Store store) {
+    public Review(Long makerId, String reviewContent, Double rate, Store store) {
         this.makerId = makerId;
         this.reviewContent = reviewContent;
         this.rate = rate;
         this.store = store;
     }
 
-    public void update(String reviewContent, Long rate) {
+    public void update(String reviewContent, Double rate) {
         this.reviewContent = reviewContent;
         this.rate = rate;
     }
